Rename modifier YAML Class key from 'compositor' to 'modifier'
Includes moving modifiers to their own satpy.modifiers subpackage
