+++ satpy/etc/composites/ahi.yaml
@@ -2,7 +2,7 @@ sensor_name: visir/ahi
 
 modifiers:
   rayleigh_corrected:
-    compositor: !!python/name:satpy.composites.PSPRayleighReflectance
+    modifier: !!python/name:satpy.modifiers.PSPRayleighReflectance
     atmosphere: us-standard
     aerosol_type: marine_clean_aerosol
     prerequisites:
