+++ satpy/etc/composites/slstr.yaml
@@ -16,7 +16,7 @@ composite_identification_keys:
 
 modifiers:
   nir_reflectance:
-    compositor: !!python/name:satpy.composites.NIRReflectance
+    modifier: !!python/name:satpy.modifiers.NIRReflectance
     prerequisites:
     - S8
     optional_prerequisites:
