+++ satpy/tests/test_dependency_tree.py
@@ -105,7 +105,8 @@ class TestMultipleResolutionSameChannelDependency(unittest.TestCase):
         from satpy.readers.yaml_reader import FileYAMLReader
 
         from satpy import DataQuery
-        from satpy.composites import SunZenithCorrector, GenericCompositor
+        from satpy.composites import GenericCompositor
+        from satpy.modifiers import SunZenithCorrector
         from satpy.dataset import DatasetDict
 
         config_file = os.path.join(PACKAGE_CONFIG_PATH, 'readers', 'modis_l1b.yaml')
